Move descriptors tests to object
